+++ GaggiaPIDController/RelayPIDController.h
@@ -6,19 +6,19 @@
 class RelayPIDController
 {
   public:
-    RelayPIDController(const uint16_t &kp, const uint16_t &ki, const uint16_t &kd);
+    RelayPIDController(const double &kp, const double &ki, const double &kd);
 
     bool compute(double *input, double *setpoint, bool *relay_on);
 
-    void set_kp(uint16_t *kp);
-    void set_ki(uint16_t *ki);
-    void set_kd(uint16_t *kd);
+    void set_kp(double *kp);
+    void set_ki(double *ki);
+    void set_kd(double *kd);
 
   private:
     PID *pid;
-    uint16_t p_gain;
-    uint16_t i_gain;
-    uint16_t d_gain;
+    double p_gain;
+    double i_gain;
+    double d_gain;
 
     double pid_input;
     double pid_output;
